# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API Documentation",
        "title": "UKAI Syndrome API V2",
        "contact": {},
        "version": "2.0"
    },
    "basePath": "/api/v2",
    "paths": {
        "/absensi/peserta/me": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get seluruh riwayat absensi peserta yang sedang login",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Absensi"
                ],
                "summary": "Get absensi peserta",
                "responses": {}
            }
        },
        "/absensi/peserta/me/check-in": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Peserta melakukan check-in pada jadwal ketika mentor sedang mengajar dan peserta berada dalam radius lokasi mentor",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Absensi"
                ],
                "summary": "Check-in peserta",
                "parameters": [
                    {
                        "description": "Data check-in peserta",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/absensi.CheckInPesertaRequest"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/absensi/peserta/me/jadwal/{id_jadwal}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Mengecek apakah peserta yang sedang login sudah melakukan check-in pada jadwal tertentu",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Absensi"
                ],
                "summary": "Get status absensi peserta berdasarkan jadwal",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID Jadwal",
                        "name": "id_jadwal",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/absensi/peserta/me/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get detail absensi milik peserta yang sedang login",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Absensi"
                ],
                "summary": "Get detail absensi peserta",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID Absensi Peserta",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
...
    },
    "definitions": {
        "absensi.CheckInPesertaRequest": {
            "type": "object",
            "properties": {
                "id_jadwal": {
                    "type": "integer"
                },
                "latitude": {
                    "type": "number"
                },
                "location_accuracy": {
                    "type": "number"
                },
                "longitude": {
                    "type": "number"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}